+++ iam-login-service/src/main/java/it/infn/mw/iam/config/oidc/IamOidcJITAccountProvisioningProperties.java
@@ -15,53 +15,13 @@
  */
 package it.infn.mw.iam.config.oidc;
 
-import static java.lang.Boolean.FALSE;
-
-import java.util.Optional;
-import java.util.Set;
-
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
-import com.google.common.base.Splitter;
-import com.google.common.collect.Sets;
+import it.infn.mw.iam.config.JitProvisioningProperties;
 
 @Validated
 @ConfigurationProperties(prefix = "oidc.jit-account-provisioning")
-public class IamOidcJITAccountProvisioningProperties {
+public class IamOidcJITAccountProvisioningProperties extends JitProvisioningProperties {
 
-  private boolean enabled = FALSE;
-  private String trustedIdps = "all";
-
-  public Boolean getEnabled() {
-    return enabled;
-  }
-
-  public void setEnabled(boolean enabled) {
-    this.enabled = enabled;
-  }
-
-  public String getTrustedIdps() {
-    return trustedIdps;
-  }
-
-  public void setTrustedIdps(String trustedIdps) {
-    this.trustedIdps = trustedIdps;
-  }
-
-  public Optional<Set<String>> getTrustedIdpsAsOptionalSet() {
-    if ("all".equals(trustedIdps)) {
-      return Optional.empty();
-    }
-
-    Set<String> trustedIdpIds =
-        Sets.newHashSet(Splitter.on(",").trimResults().omitEmptyStrings().split(trustedIdps));
-
-    if (trustedIdpIds.isEmpty()) {
-      return Optional.empty();
-    }
-
-    return Optional.of(trustedIdpIds);
-  }
 }
-
